Improvement: --going will fix all symlinks before telling you it's now safe to disconnect your drive
For shares with only 1 file copy, those files will be copied somewhere
else first. (This was already implemented.)
For shares with more copies, Greyhole will now look, in the shared
folder (landing zone), for symlinks that point to the soon-to-been-gone
drive, and will change all of them to point to another valid copy
instead.
Closes GH-5