Clarification on what trash does
Trash does not just keep files that have been deleted. Files that have been modified are also saved in trash as a revision backup for scenarios when a user may of made a mistake and wishes to revert. Or in the hopefully unlikely event that greyhole does something it was not meant to have done.